Handle reserved keywords
[Similarly to apache/thrift](https://github.com/apache/thrift/blob/19baeefd8c38d62085891d7956349601f79448b3/compiler/cpp/src/thrift/thriftl.ll#L273), we disallow keywords being used in the Thrift IDL
However the keywords restriction only applies to exact-casing, so for example we must support "AND" as a valid enum. When down-casing this value, it conflicts with the Elixir keyword "and", and so using some macro magic we can get around this limitation.
Additionally, since we down-case method names too, similar logic applies when generating a client. The server implementation did not require any changes.

Fixes: #294

+++ test/thrift/parser/file_group_test.exs
@@ -8,7 +8,7 @@ defmodule Thrift.Parser.FileGroupTest do
   test "constant module uses suitable existing name" do
     with_thrift_files([
       "myservice.thrift": """
-      const float PI = 3.14
+      const double PI = 3.14
       service MyService {}
       """, as: :file_group, parse: "myservice.thrift"]) do
